requirements.txt
python-telegram-bot
aiosqlite
python-dotenv
nest_asyncio
pytz